Folder refactor and added env template
