+++ sports-scores/.env.template
@@ -1,0 +1,7 @@
+# API Keys
+SportsMonksAPIKey=
+CricketDataAPI=
+APISportsKey=
+
+# Base URLs
+AFL_BASEURL=
