+++ sports-scores/src/components/ui/NavButtonGroup.tsx
