+++ sports-scores/src/lib/constants.ts
