+++ sports-scores/src/lib/types.ts
